CmbDB: Add support for save-always properties
We can not always rely on a property default value to know if it should
be serialized or not, since some properties value change at runtime
depending when or how the widget was created.

For example GtkGrid children property change depending when the child was added so
column and row should always be serialized.

GtkGrid data: mark column, row, left-attach, top-attach as save always.

Closses issue #10