CmbApplication: improve file open handling
Cambalache is associated with .cmb, .ui and .glade files but
it only opened projects files (.cmb) you had to manually import the others.

This add support for importing UI files in the current project or
create a new one if needed when trying to open a file.

If the file is already imported in the project, it will be selected.

Closes issue #45